+++ jqm-all/jqm-integration-tests/src/test/java/com/enioka/jqm/integration/tests/ApiSimpleTest.java
@@ -99,10 +99,10 @@ public class ApiSimpleTest extends JqmBaseTest
 
         String result = res.body();
 
-        Integer jid = 0;
+        long jid = 0L;
         try
         {
-            jid = Integer.parseInt(result);
+            jid = Long.parseLong(result);
         }
         catch (Exception e)
         {
@@ -148,7 +148,7 @@ public class ApiSimpleTest extends JqmBaseTest
 
         String result = res.body();
 
-        Integer jid = 0;
+        long jid = 0;
         try
         {
             jid = Integer.parseInt(result);
